Reorganizing components before redesign.
