+++ src/buttons/SubtractHpButton.js
